--- .lintstagedrc.json
@@ -1,3 +1,0 @@
-{
-  "*.js": "eslint"
-}
